pyproject.toml
[project]
name = "project-codeguard"
version = "1.3.1"
description = "AI Coding Rules for Security and Best Practices"
requires-python = ">=3.11"
dependencies = [
    "pyyaml~=6.0.2",
    "mkdocs~=1.6.1",
    "mkdocs-material~=9.6.21",
]

[project.optional-dependencies]
benchmarks = [
    "docker~=7.1",
    "httpx~=0.28",
    "pydantic~=2.11",
    "python-dotenv~=1.1",
]

[tool.uv]
# CVE floors for transitive deps
constraint-dependencies = [
    "urllib3>=2.6.3",
    "requests>=2.33.0",
    "pygments>=2.20.0",
]
# Supply-chain cooldown
exclude-newer = "7 days"
